Automatically choose dictionary size on DOS hosts
LZMA uses an amount of XMS memory that is approximately thirteen times
the size of the encoding dictionary, plus the size of the static
allocation of the 7z executable image.

None of the upstream compression presets comfortably fit the intended
DOS target machine. Rounding up:

  -mx1 requires 8 MiB
  -mx2 requires 16 MiB
  -mx3 requires 64 MiB
  -mx4 requires 256 MiB
  -mx5 requires 512 MiB (this is the upstream default)

If a dictionary size is not passed with a -md switch, then override the
working preset and just use the largest dictionary that can be allocated
by the host.

This means that a 64 MiB host will create 7z files with a 2 MiB
dictionary that can be unpacked by the 7zm mini variant running on a 6
MiB host or the 7zr reduced variant running on a 8 MiB host.