Improve the Win32c console UX.
The welcome text in 7-Zip confuses new and casual users. Rewrite it so
that it fits on one 80x25 console screen and captures 95% of what people
do with 7z on the command line.

Rewrite --help so that it looks more like a man page, is easier to read,
and is more informative.

Put the git commit in the program banner and move the copyright
attribution to a new --version switch.