Resolve compiler warning 014
W014: no reference to symbol

The pthread_dummy.h header will never be implemented for DOS, so disable
warnings there with a pramga.

The deflate codec is internally reused by some modules in a way that causes
warnings for unreferenced constants, so disable warnings there too with a
pragma.

Wrap the DPMI memory struct with a pragma because all of its fields are
mandatory for the int386() call but most of its fields are unreferenced.

Disable PrintNum(), PrintStat(), and PrintTime() because those functions will
be unused until the benchmark command is implmented.

Disable GetRelativePath() in DOS builds because it is used only for handling
links, which are not implemented by FAT filesystems.

Delete FiTime_To_timespec() in the DOS platform code because it is unused.

Closes: #25